+++ api/proto/v1/auditoria_servicio.pb.gw.go
@@ -35,33 +35,6 @@ var (
 	_ = metadata.Join
 )
 
-func request_AuditoriaService_RegistrarEvento_0(ctx context.Context, marshaler runtime.Marshaler, client AuditoriaServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
-	var (
-		protoReq RegistrarEventoRequest
-		metadata runtime.ServerMetadata
-	)
-	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && !errors.Is(err, io.EOF) {
-		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
-	}
-	if req.Body != nil {
-		_, _ = io.Copy(io.Discard, req.Body)
-	}
-	msg, err := client.RegistrarEvento(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
-	return msg, metadata, err
-}
-
-func local_request_AuditoriaService_RegistrarEvento_0(ctx context.Context, marshaler runtime.Marshaler, server AuditoriaServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
-	var (
-		protoReq RegistrarEventoRequest
-		metadata runtime.ServerMetadata
-	)
-	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && !errors.Is(err, io.EOF) {
-		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
-	}
-	msg, err := server.RegistrarEvento(ctx, &protoReq)
-	return msg, metadata, err
-}
-
 var filter_AuditoriaService_ConsultarEventos_0 = &utilities.DoubleArray{Encoding: map[string]int{}, Base: []int(nil), Check: []int(nil)}
 
 func request_AuditoriaService_ConsultarEventos_0(ctx context.Context, marshaler runtime.Marshaler, client AuditoriaServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
@@ -103,26 +76,6 @@ func local_request_AuditoriaService_ConsultarEventos_0(ctx context.Context, mars
 // Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterAuditoriaServiceHandlerFromEndpoint instead.
 // GRPC interceptors will not work for this type of registration. To use interceptors, you must use the "runtime.WithMiddlewares" option in the "runtime.NewServeMux" call.
 func RegisterAuditoriaServiceHandlerServer(ctx context.Context, mux *runtime.ServeMux, server AuditoriaServiceServer) error {
-	mux.Handle(http.MethodPost, pattern_AuditoriaService_RegistrarEvento_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
-		ctx, cancel := context.WithCancel(req.Context())
-		defer cancel()
-		var stream runtime.ServerTransportStream
-		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
-		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		annotatedContext, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/dbgs.v1.AuditoriaService/RegistrarEvento", runtime.WithHTTPPathPattern("/v1/auditoria/eventos"))
-		if err != nil {
-			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
-			return
-		}
-		resp, md, err := local_request_AuditoriaService_RegistrarEvento_0(annotatedContext, inboundMarshaler, server, req, pathParams)
-		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
-		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
-		if err != nil {
-			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
-			return
-		}
-		forward_AuditoriaService_RegistrarEvento_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
-	})
 	mux.Handle(http.MethodGet, pattern_AuditoriaService_ConsultarEventos_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
@@ -183,23 +136,6 @@ func RegisterAuditoriaServiceHandler(ctx context.Context, mux *runtime.ServeMux,
 // doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
 // "AuditoriaServiceClient" to call the correct interceptors. This client ignores the HTTP middlewares.
 func RegisterAuditoriaServiceHandlerClient(ctx context.Context, mux *runtime.ServeMux, client AuditoriaServiceClient) error {
-	mux.Handle(http.MethodPost, pattern_AuditoriaService_RegistrarEvento_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
-		ctx, cancel := context.WithCancel(req.Context())
-		defer cancel()
-		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		annotatedContext, err := runtime.AnnotateContext(ctx, mux, req, "/dbgs.v1.AuditoriaService/RegistrarEvento", runtime.WithHTTPPathPattern("/v1/auditoria/eventos"))
-		if err != nil {
-			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
-			return
-		}
-		resp, md, err := request_AuditoriaService_RegistrarEvento_0(annotatedContext, inboundMarshaler, client, req, pathParams)
-		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
-		if err != nil {
-			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
-			return
-		}
-		forward_AuditoriaService_RegistrarEvento_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)
-	})
 	mux.Handle(http.MethodGet, pattern_AuditoriaService_ConsultarEventos_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
@@ -221,11 +157,9 @@ func RegisterAuditoriaServiceHandlerClient(ctx context.Context, mux *runtime.Ser
 }
 
 var (
-	pattern_AuditoriaService_RegistrarEvento_0  = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2}, []string{"v1", "auditoria", "eventos"}, ""))
 	pattern_AuditoriaService_ConsultarEventos_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2}, []string{"v1", "auditoria", "eventos"}, ""))
 )
 
 var (
-	forward_AuditoriaService_RegistrarEvento_0  = runtime.ForwardResponseMessage
 	forward_AuditoriaService_ConsultarEventos_0 = runtime.ForwardResponseMessage
 )